+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Added `Resampler::reset` method for resetting internal state
 - Enhanced documentation of `Resampler::finalize` to mention need to
   loop